+++ src/main/java/de/letsdev/maven/plugins/ios/ProvisioningProfileHelper.java
@@ -1,5 +1,17 @@
+/**
+ * ios-maven-plugin
+ * <p/>
+ * User: mauer
+ * Date: 2019-07-09
+ * <p/>
+ * This code is copyright (c) 2019 let's dev.
+ * URL: https://www.letsdev.de
+ * e-Mail: contact@letsdev.de
+ */
+
 package de.letsdev.maven.plugins.ios;
 
+import de.letsdev.maven.plugins.ios.mojo.BaseMojo;
 import de.letsdev.maven.plugins.ios.mojo.IOSException;
 import org.apache.maven.project.MavenProject;
 import org.w3c.dom.Document;
@@ -14,6 +26,17 @@ import java.util.Map;
 
 public class ProvisioningProfileHelper {
 
+    private static final String NODE_NAME_STRING = "string";
+    private static final String NODE_NAME_ARRAY = "array";
+    private static final String NODE_NAME_KEY = "key";
+    private static final String NODE_NAME_TRUE = "true";
+    private static final String TEAM_ID_KEY = "TeamIdentifier";
+    private static final String UUID_KEY = "UUID";
+    private static final String NAME_KEY = "Name";
+    private static final String APP_ID_KEY = "application-identifier";
+    private static final String GET_TASK_ALLOW_KEY = "get-task-allow";
+    private static final String PROVISION_DEVICES_KEY = "ProvisionsAllDevices";
+    private static final String PROVISIONED_DEVICES_KEY = "ProvisionedDevices";
     private static final String provisioningProfileDirectory =
             System.getProperty("user.home") + "/Library/MobileDevice/Provisioning\\ " + "Profiles";
     private static final String provisioningProfileFileExtension = ".mobileprovision";
@@ -68,28 +91,35 @@ public class ProvisioningProfileHelper {
             doc = dBuilder.parse(inputFile);
             doc.getDocumentElement().normalize();
             String uuid = null;
-            String teamID = null;
+            String teamId = null;
             String name = null;
+            String bundleId = null;
 
-            NodeList nodeList = doc.getElementsByTagName("key");
+            NodeList nodeList = doc.getElementsByTagName(NODE_NAME_KEY);
             for (int i = 0; i < nodeList.getLength(); i++) {
-                if (nodeList.item(i).getFirstChild().getNodeValue().equals("UUID")) {
+                if (nodeList.item(i).getFirstChild().getNodeValue().equals(UUID_KEY)) {
                     uuid = getStringValue(nodeList, i);
-            }
+                }
 
-                if (nodeList.item(i).getFirstChild().getNodeValue().equals("Name")) {
+                if (nodeList.item(i).getFirstChild().getNodeValue().equals(NAME_KEY)) {
                     name = getStringValue(nodeList, i);
                 }
 
-                if (nodeList.item(i).getFirstChild().getNodeValue().equals("TeamIdentifier")) {
-                    teamID = getTeamId(nodeList, i);
+                if (nodeList.item(i).getFirstChild().getNodeValue().equals(TEAM_ID_KEY)) {
+                    teamId = getTeamId(nodeList, i);
+                }
+
+                if (nodeList.item(i).getFirstChild().getNodeValue().equals(APP_ID_KEY)) {
+                    bundleId = getStringValue(nodeList, i);
                 }
             }
 
             ProvisioningProfileType type = getProvisioningProfileType(nodeList);
 
-            if (teamID != null && uuid != null && name != null) {
-                return new ProvisioningProfileData(uuid, name, teamID, type);
+            if (teamId != null && uuid != null && name != null) {
+                bundleId = removeTeamFromBundleId(bundleId, teamId);
+
+                return new ProvisioningProfileData(uuid, name, teamId, bundleId, type);
             }
         } catch (Exception e) {
             e.printStackTrace();
@@ -98,26 +128,36 @@ public class ProvisioningProfileHelper {
         return null;
     }
 
-    private String getStringValue(NodeList nodeList, int i) {
-            Node current = nodeList.item(i).getNextSibling();
-            //iterate trough notes until node with value for the uuid key is current
-            while (!current.getNodeName().equals("string")) {
-                current = current.getNextSibling();
-            }
+    private String removeTeamFromBundleId(String bundleId, String teamId) {
+        if (bundleId != null && !bundleId.isEmpty()) {
+            bundleId = bundleId.replaceAll(teamId + ".", "");
+        }
 
-        return  current.getFirstChild().getNodeValue();
+        return bundleId;
+    }
+
+    private String getStringValue(NodeList nodeList, int i) {
+
+        Node current = nodeList.item(i).getNextSibling();
+        //iterate trough notes until node with value for the uuid key is current
+        while (!current.getNodeName().equals(NODE_NAME_STRING)) {
+            current = current.getNextSibling();
+        }
+
+        return current.getFirstChild().getNodeValue();
     }
 
     private String getTeamId(NodeList nodeList, int i) {
-            Node current = nodeList.item(i).getNextSibling();
-            //iterate trough notes until array node, that has value of team id key as child is current
-            while (!current.getNodeName().equals("array")) {
-                current = current.getNextSibling();
-            }
-            current = current.getFirstChild();
-            while (!current.getNodeName().equals("string")) {
-                current = current.getNextSibling();
-            }
+
+        Node current = nodeList.item(i).getNextSibling();
+        //iterate trough notes until array node, that has value of team id key as child is current
+        while (!current.getNodeName().equals(NODE_NAME_ARRAY)) {
+            current = current.getNextSibling();
+        }
+        current = current.getFirstChild();
+        while (!current.getNodeName().equals(NODE_NAME_STRING)) {
+            current = current.getNextSibling();
+        }
         return current.getFirstChild().getNodeValue();
     }
 
@@ -131,17 +171,17 @@ public class ProvisioningProfileHelper {
         ProvisioningProfileType type = null;
         for (int i = 0; i < nodeList.getLength(); i++) {
 
-            if (checkKey("get-task-allow", nodeList.item(i))) {
+            if (checkKey(GET_TASK_ALLOW_KEY, nodeList.item(i))) {
                 type = ProvisioningProfileType.TYPE_DEVELOPMENT;
                 break;
-            } else if (checkKey("ProvisionsAllDevices", nodeList.item(i))) {
+            } else if (checkKey(PROVISION_DEVICES_KEY, nodeList.item(i))) {
                 type = ProvisioningProfileType.TYPE_ENTERPRISE;
                 break;
-            } else if (nodeList.item(i).getFirstChild().getNodeValue().equals("ProvisionedDevices")) {
+            } else if (nodeList.item(i).getFirstChild().getNodeValue().equals(PROVISIONED_DEVICES_KEY)) {
                 Node current = nodeList.item(i).getNextSibling();
                 while (true) {
                     current = current.getNextSibling();
-                    if (current.getNodeName().equals("array")) {
+                    if (current.getNodeName().equals(NODE_NAME_ARRAY)) {
                         if (current.hasChildNodes()) {
                             type = ProvisioningProfileType.TYPE_AD_HOC;
                         }
@@ -149,7 +189,7 @@ public class ProvisioningProfileHelper {
                         break;
                     }
 
-                    if (current.getNodeName().equals("key")) {
+                    if (current.getNodeName().equals(NODE_NAME_KEY)) {
                         break;
                     }
                 }
@@ -168,10 +208,10 @@ public class ProvisioningProfileHelper {
             Node current = node.getNextSibling();
             while (true) {
                 current = current.getNextSibling();
-                if (current.getNodeName().equals("true")) {
+                if (current.getNodeName().equals(NODE_NAME_TRUE)) {
                     return true;
                 }
-                if (current.getNodeName().equals("key")) {
+                if (current.getNodeName().equals(NODE_NAME_KEY)) {
                     return false;
                 }
             }