+++ src/main/java/de/letsdev/maven/plugins/ios/mojo/BaseMojo.java
@@ -567,14 +567,20 @@ public class BaseMojo extends AbstractMojo {
                 ProvisioningProfileData data = helper.getData();
 
                 if (data != null) {
+
+                    if (data.getBundleId() != null && !data.getBundleId().contains("*")) {
+                        this.bundleIdentifier = data.getBundleId();
+                    }
+
+                    this.provisioningProfileUUID = data.getUuid();
+                    this.provisioningProfileSpecifier = data.getName();
+                    this.developmentTeam = data.getTeamId();
+
                     if (this.xcodeExportOptions.provisioningProfiles == null) {
                         this.xcodeExportOptions.provisioningProfiles = new HashMap<>();
                     }
                     this.xcodeExportOptions.provisioningProfiles.put(this.bundleIdentifier, data.getUuid());
-                    this.provisioningProfileUUID = data.getUuid();
-                    this.provisioningProfileSpecifier = data.getName();
-                    this.xcodeExportOptions.teamID = data.getTeamID();
-                    this.developmentTeam = data.getTeamID();
+                    this.xcodeExportOptions.teamID = data.getTeamId();
                     this.xcodeExportOptions.method = data.getTypeId();
 
                     this.properties = prepareProperties();