+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+### ADDED
+- bundle identifier determination from .mobileprovision file
+
 ## [1.18.11] - 2019-12-17
 ### Added
 - xcframeworks will be generated per default (iOSXcFrameworkBuild), when building a iOS/macOS library